pyproject.toml
[tool.poetry]
name = "rd_assistant"
version = "0.1.0"
description = "AI-powered requirements gathering assistant"
authors = ["Makoto HAMATSU <@hamatz>"]

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.12.0"
anthropic = "^0.18.1"
python-dotenv = "^1.0.0"
rich = "^13.7.0"
prompt-toolkit = "^3.0.43"

[tool.poetry.dev-dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"